Make it easier to add more perf benchmarks
