+++ test/Channels.Tests.Performance/Channels.Tests.Performance.xproj
@@ -7,7 +7,7 @@
   <Import Project="$(VSToolsPath)\DotNet\Microsoft.DotNet.Props" Condition="'$(VSToolsPath)' != ''" />
   <PropertyGroup Label="Globals">
     <ProjectGuid>fd7d92ee-542c-424b-b66a-9ac2a214ec90</ProjectGuid>
-    <RootNamespace>Channels.Performance</RootNamespace>
+    <RootNamespace>Channels.Tests.Performance</RootNamespace>
     <BaseIntermediateOutputPath Condition="'$(BaseIntermediateOutputPath)'=='' ">.\obj</BaseIntermediateOutputPath>
     <OutputPath Condition="'$(OutputPath)'=='' ">.\bin\</OutputPath>
     <TargetFrameworkVersion>v4.6.2</TargetFrameworkVersion>
