+++ test/Channels.Tests.Performance/Program.cs
@@ -1,36 +1,52 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
-using System.IO.Compression;
-using System.Text.Utf8;
-using System.Threading.Tasks;
-using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Configs;
 using BenchmarkDotNet.Diagnostics.Windows;
 using BenchmarkDotNet.Jobs;
 using BenchmarkDotNet.Running;
-using Newtonsoft.Json;
-using Channels.Compression;
-
-using JsonReader = System.Text.Json.JsonReader;
-using Channels.File;
 
 namespace Channels.Tests.Performance
 {
     public class Program
     {
-        public static void Main()
+        public static void Main(string[] args)
         {
-            BenchmarkRunner.Run<ChannelsBenchmark>();
+            var options = (uint[])Enum.GetValues(typeof(BenchmarkType));
+            BenchmarkType type;
+            if (args.Length != 1 || !Enum.TryParse(args[0], out type))
+            {
+                Console.WriteLine($"Please add benchmark to run as parameter:");
+                for (var i = 0; i < options.Length; i++)
+                {
+                    Console.WriteLine($"  {((BenchmarkType)options[i]).ToString()}" );
+                }
 
-            //ChannelsBenchmark.Streams();
+                return;
+            }
+
+            RunSelectedBenchmarks(type);
+        }
+
+        private static void RunSelectedBenchmarks(BenchmarkType type)
+        {
+            if (type.HasFlag(BenchmarkType.Streams))
+            {
+                BenchmarkRunner.Run<ChannelsStreamsBenchmark>();
+            }
         }
     }
 
-    public class Config : ManualConfig
+    [Flags]
+    public enum BenchmarkType : uint
     {
-        public Config()
+        Streams = 1,
+        // add new ones in powers of two - e.g. 2,4,8,16...
+
+        All = uint.MaxValue
+    }
+
+    public class DefaultConfig : ManualConfig
+    {
+        public DefaultConfig()
         {
             Add(Job.Default.
                 With(Platform.X64).
@@ -40,182 +56,8 @@ namespace Channels.Tests.Performance
                 WithIterationTime(200). // 200ms per iteration
                 WithWarmupCount(5).
                 WithTargetCount(10));
+
             Add(new MemoryDiagnoser());
-        }
-    }
-
-    [Config(typeof(Config))]
-    public class ChannelsBenchmark
-    {
-        public const int InnerLoopCount = 50;
-
-        public static MemoryPool Pool;
-        public static ChannelFactory ChannelFactory;
-
-        [Setup]
-        public static void Setup()
-        {
-            if (Pool == null)
-            {
-                Pool = new MemoryPool();
-                Pool.Lease(1024).Dispose();
-            }
-
-            if (ChannelFactory == null)
-            {
-                ChannelFactory = new ChannelFactory(Pool);
-            }
-            
-        }
-
-        [Benchmark(OperationsPerInvoke = InnerLoopCount)]
-        public static void Channels()
-        {
-            ChannelsAsync().Wait();
-        }
-
-        private static async Task ChannelsAsync()
-        {
-            var cf = ChannelFactory;
-            for (var i = 0; i < InnerLoopCount; i++)
-            {
-                var input = cf.ReadFile("random.json.gz")
-                                .GZipDecompress(cf)
-                                .GZipCompress(cf, CompressionLevel.NoCompression)
-                                .GZipDecompress(cf)
-                                .GZipCompress(cf, CompressionLevel.NoCompression)
-                                .GZipDecompress(cf)
-                                .GZipCompress(cf, CompressionLevel.NoCompression)
-                                .GZipDecompress(cf)
-                                .GZipCompress(cf, CompressionLevel.NoCompression)
-                                .GZipDecompress(cf)
-                                .GZipCompress(cf, CompressionLevel.NoCompression)
-                                .GZipDecompress(cf);
-
-                var data = await input.ReadToEndAsync();
-
-                //ParseJson(data);
-                input.Advance(data.End);
-            }
-        }
-
-
-        private static unsafe void ParseJson(ReadableBuffer buffer)
-        {
-            Utf8String json;
-            if (buffer.IsSingleSpan) {
-                json = new Utf8String(buffer.First.Span);
-            }
-            else {
-                var length = buffer.Length;
-                byte* b = stackalloc byte[length];
-                buffer.CopyTo(new Span<byte>(b, length));
-                json = new Utf8String(new ReadOnlySpan<byte>(b, length));
-            }
-            //Console.WriteLine(json.Length);
-
-            var reader = new JsonReader(json);
-
-            while (reader.Read())
-            {
-                var tokenType = reader.TokenType;
-                switch (tokenType)
-                {
-                    case JsonReader.JsonTokenType.ObjectStart:
-                    case JsonReader.JsonTokenType.ObjectEnd:
-                    case JsonReader.JsonTokenType.ArrayStart:
-                    case JsonReader.JsonTokenType.ArrayEnd:
-                        //Console.WriteLine(tokenType);
-                        break;
-                    case JsonReader.JsonTokenType.Property:
-                        var name = reader.GetName();
-                        //Console.WriteLine(name);
-                        var value = reader.GetValue();
-                        //Console.WriteLine(value);
-                        break;
-                    case JsonReader.JsonTokenType.Value:
-                        value = reader.GetValue();
-                        //Console.WriteLine(value);
-                        break;
-                    default:
-                        throw new ArgumentOutOfRangeException();
-                }
-            }
-        }
-
-        [Benchmark(Baseline = true, OperationsPerInvoke = InnerLoopCount)]
-        public static void Streams()
-        {
-            StreamsAsync().Wait();
-        }
-
-
-        public static async Task StreamsAsync()
-        {
-            for (var i = 0; i < InnerLoopCount; i++)
-            {
-                var mem1 = new MemoryStream();
-
-                using (var filestream = new FileStream("random.json.gz", FileMode.Open, FileAccess.Read))
-                using (var gzip0 = new GZipStream(filestream, CompressionMode.Decompress))
-                {
-                    using (var gzip1 = new GZipStream(mem1, CompressionLevel.NoCompression, true))
-                    {
-                        await gzip0.CopyToAsync(gzip1);
-                    }
-                }
-                mem1.Seek(0, SeekOrigin.Begin);
-
-                var mem3 = new MemoryStream();
-                using (var gzip2 = new GZipStream(mem1, CompressionMode.Decompress, true))
-                {
-                    using (var gzip3 = new GZipStream(mem3, CompressionLevel.NoCompression, true))
-                    {
-                        await gzip2.CopyToAsync(gzip3);
-                    }
-                }
-                mem3.Seek(0, SeekOrigin.Begin);
-
-                var mem5 = new MemoryStream();
-                using (var gzip4 = new GZipStream(mem3, CompressionMode.Decompress, true))
-                {
-                    using (var gzip5 = new GZipStream(mem5, CompressionLevel.NoCompression, true))
-                    {
-                        await gzip4.CopyToAsync(gzip5);
-                    }
-                }
-                mem5.Seek(0, SeekOrigin.Begin);
-
-                var mem7 = new MemoryStream();
-                using (var gzip6 = new GZipStream(mem5, CompressionMode.Decompress, true))
-                {
-                    using (var gzip7 = new GZipStream(mem7, CompressionLevel.NoCompression, true))
-                    {
-                        await gzip6.CopyToAsync(gzip7);
-                    }
-                }
-                mem7.Seek(0, SeekOrigin.Begin);
-
-                var mem9 = new MemoryStream();
-                using (var gzip8 = new GZipStream(mem7, CompressionMode.Decompress, true))
-                {
-                    using (var gzip9 = new GZipStream(mem9, CompressionLevel.NoCompression, true))
-                    {
-                        await gzip8.CopyToAsync(gzip9);
-                    }
-                }
-                mem9.Seek(0, SeekOrigin.Begin);
-
-                using (var gzip10 = new GZipStream(mem9, CompressionMode.Decompress, true))
-                {
-                    using (var textReader = new StreamReader(gzip10))
-                    {
-                        var data = await textReader.ReadToEndAsync();
-
-                        // Do stuff with data
-                    }
-                }
-            }
         }
     }
 }
